@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.15)
 project(velesquant LANGUAGES CXX)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if(MSVC)
