@@ -1,11 +1,9 @@
 #include <velesquant/models/quantoed_cms.h>
 
-#include <algorithm>
-#include <string>
 #include <vector>
-#include <velesquant/volatility/sabr.h>
 #include <velesquant/models/utility.h>
 #include <velesquant/types.h>
+#include <velesquant/volatility/sabr.h>
 
 using namespace std;
 using namespace velesquant::xlw;
@@ -13,9 +11,10 @@ using namespace velesquant::xlw;
 namespace velesquant {
 
 quantoedCMS::quantoedCMS(double expirySR, double tenorSR, double forwardSR,
-                         double annuitySR, double payCMS, double discountCMS,
-                         double corFX, double atmVolFX, double beta,
-                         CellMatrix &strikes, CellMatrix &marketQuotes,
+                         double annuitySR, [[maybe_unused]] double payCMS,
+                         double discountCMS, double corFX, double atmVolFX,
+                         double beta, CellMatrix &strikes,
+                         CellMatrix &marketQuotes,
                          CalibrationTarget quoteType) {
   Sabr *swaptionSABR = new Sabr(expirySR, forwardSR, beta);
   int m = strikes.RowsInStructure();
