@@ -15,10 +15,8 @@
 #include <velesquant/models/solver.h>
 #include <velesquant/numerics/interpolation.h>
 
-using namespace std;
-using namespace QuantLib;
-
-#pragma warning(disable : 4996)
+// using namespace std;
+// using namespace QuantLib;
 
 namespace velesquant {
 
@@ -47,7 +45,7 @@ double implied_vol2(double Maturity, double Forward, double Strike,
 };
 
 double annuity(double Expiry, double Tenor, double Freq,
-               const vector<double> &DF, const vector<double> &T) {
+               const std::vector<double> &DF, const std::vector<double> &T) {
   double ti, DFi;
   double annuity = 0.0;
 
@@ -61,8 +59,8 @@ double annuity(double Expiry, double Tenor, double Freq,
   return annuity;
 };
 
-double fwdSR(double Expiry, double Tenor, double Freq, const vector<double> &DF,
-             const vector<double> &T) {
+double fwdSR(double Expiry, double Tenor, double Freq,
+             const std::vector<double> &DF, const std::vector<double> &T) {
   double D0, DN;
 
   D0 = interpolation("LinearInterpolation", T, DF, Expiry);
@@ -71,15 +69,16 @@ double fwdSR(double Expiry, double Tenor, double Freq, const vector<double> &DF,
   return (D0 - DN) / annuity(Expiry, Tenor, Freq, DF, T);
 };
 
-vector<vector<double>> cholesky(const vector<vector<double>> &corrMatrix) {
+std::vector<std::vector<double>>
+cholesky(const std::vector<std::vector<double>> &corrMatrix) {
   int N = int(corrMatrix.size());
-  Matrix AAA(N, N);
+  QuantLib::Matrix AAA(N, N);
   for (int i = 0; i < N; i++)
     for (int j = 0; j < N; j++)
       AAA[i][j] = corrMatrix[i][j];
-  Matrix BBB(N, N);
-  BBB = CholeskyDecomposition(AAA);
-  vector<vector<double>> matrixCholesky(N, vector<double>(N));
+  QuantLib::Matrix BBB(N, N);
+  BBB = QuantLib::CholeskyDecomposition(AAA);
+  std::vector<std::vector<double>> matrixCholesky(N, std::vector<double>(N));
   // matrixCholesky.resize(N);
   for (int i = 0; i < N; i++) {
     //		matrixCholesky[i].resize(N);
@@ -90,13 +89,13 @@ vector<vector<double>> cholesky(const vector<vector<double>> &corrMatrix) {
 };
 
 double option_vega(double F, double K, double sigma, double T) {
-  double sig = sigma * sqrt(T);
-  double d1 = (log(F / K) + sig * sig) / sig;
+  double sig = sigma * std::sqrt(T);
+  double d1 = (std::log(F / K) + sig * sig) / sig;
   return F * pdf_normal(d1);
 };
 
-void lAxy(const vector<vector<double>> &lowerTriangular,
-          const vector<double> &input, vector<double> &result) {
+void lAxy(const std::vector<std::vector<double>> &lowerTriangular,
+          const std::vector<double> &input, std::vector<double> &result) {
   int n = input.size();
   for (int i = 0; i < n; i++) {
     result[i] = 0;
@@ -106,26 +105,29 @@ void lAxy(const vector<vector<double>> &lowerTriangular,
 };
 
 // Converts DF curve to spot rates
-vector<double> DFtoR(const vector<double> &DF, const vector<double> &T, int n) {
+std::vector<double> DFtoR(const std::vector<double> &DF,
+                          const std::vector<double> &T, int n) {
   int m = DF.size();
-  vector<double> R0s(m);
+  std::vector<double> R0s(m);
 
   for (int i = 0; i < m; i++) {
     if (n == 0) {
-      R0s[i] = -log(DF[i]) / T[i];
+      R0s[i] = -std::log(DF[i]) / T[i];
     } else
-      R0s[i] = n * (pow(1.0 / DF[i], 1 / (n * T[i])) - 1.0);
+      R0s[i] = n * (std::pow(1.0 / DF[i], 1 / (n * T[i])) - 1.0);
   }
 
   return R0s;
 };
 
 // Calculates Div yield based on forward prices, spot and DF curve
-vector<double> DFtoDiv(const vector<double> &DF, const vector<double> &dfT,
-                       const vector<double> &Fwd, const vector<double> &fwdT,
-                       double Spot, int n) {
+std::vector<double> DFtoDiv(const std::vector<double> &DF,
+                            const std::vector<double> &dfT,
+                            const std::vector<double> &Fwd,
+                            const std::vector<double> &fwdT, double Spot,
+                            int n) {
   int m = Fwd.size();
-  vector<double> Divs(m), R0s(m);
+  std::vector<double> Divs(m), R0s(m);
 
   double interpR0;
 
@@ -134,17 +136,17 @@ vector<double> DFtoDiv(const vector<double> &DF, const vector<double> &dfT,
 
   for (int i = 0; i < m; i++) {
     interpR0 = interp(fwdT[i]);
-    Divs[i] = interpR0 - log(Fwd[i] / Spot) / fwdT[i];
+    Divs[i] = interpR0 - std::log(Fwd[i] / Spot) / fwdT[i];
   }
 
   return Divs;
 };
 
 // Calculates forward rates based on the DF curve
-vector<double> DFtoFwd(const vector<double> &DF, const vector<double> &T,
-                       double delta, int n) {
+std::vector<double> DFtoFwd(const std::vector<double> &DF,
+                            const std::vector<double> &T, double delta, int n) {
   int m = DF.size();
-  vector<double> Fwds(m), R1(m);
+  std::vector<double> Fwds(m), R1(m);
   double R2, F1, F2;
 
   R1 = DFtoR(DF, T, n);
@@ -157,9 +159,9 @@ vector<double> DFtoFwd(const vector<double> &DF, const vector<double> &T,
     if (n == 0) {
       Fwds[i] = (R2 * (T[i] + delta) - R1[i] * T[i]) / delta;
     } else {
-      F1 = pow(1.0 + R1[i] / n, n * T[i]);
-      F2 = pow(1.0 + R2 / n, n * (T[i] + delta));
-      Fwds[i] = n * (pow(F2 / F1, 1.0 / (n * delta)) - 1.0);
+      F1 = std::pow(1.0 + R1[i] / n, n * T[i]);
+      F2 = std::pow(1.0 + R2 / n, n * (T[i] + delta));
+      Fwds[i] = n * (std::pow(F2 / F1, 1.0 / (n * delta)) - 1.0);
     }
   }
 
@@ -168,19 +170,20 @@ vector<double> DFtoFwd(const vector<double> &DF, const vector<double> &T,
 
 // Calculates the forward price
 double FwdPrice(double Spot, double R0, double Div, double T) {
-  return Spot * exp((R0 - Div) * T);
+  return Spot * std::exp((R0 - Div) * T);
 };
 
-double asinh(double a) { return log(a + sqrt(a * a + 1.0)); };
+double asinh(double a) { return std::log(a + std::sqrt(a * a + 1.0)); };
 
 // Calculates the log return volatilty for a MC simulation.
 // This should converge towards the implied volatility.
-vector<double> MCimpVol(double freq, const vector<double> &theTimes,
-                        const vector<vector<double>> &thelogReturns) {
+std::vector<double>
+MCimpVol(double freq, const std::vector<double> &theTimes,
+         const std::vector<std::vector<double>> &thelogReturns) {
   int m = thelogReturns.size();
   int n = thelogReturns[0].size();
 
-  vector<double> var(m), impVol(m);
+  std::vector<double> var(m), impVol(m);
   double dt;
 
   // Initial time step
@@ -188,37 +191,37 @@ vector<double> MCimpVol(double freq, const vector<double> &theTimes,
 
   // calculate log return standard deviation
   double sum =
-      accumulate(thelogReturns[0].begin(), thelogReturns[0].end(), 0.0);
+      std::accumulate(thelogReturns[0].begin(), thelogReturns[0].end(), 0.0);
   double mean = sum / n;
   double sq_sum =
-      inner_product(thelogReturns[0].begin(), thelogReturns[0].end(),
-                    thelogReturns[0].begin(), 0.0);
-  double stDev = sqrt(sq_sum / n - mean * mean);
+      std::inner_product(thelogReturns[0].begin(), thelogReturns[0].end(),
+                         thelogReturns[0].begin(), 0.0);
+  double stDev = std::sqrt(sq_sum / n - mean * mean);
 
   // calculates implied volatility
-  double fwdVol = stDev * sqrt(freq);
+  double fwdVol = stDev * std::sqrt(freq);
 
   var[0] = fwdVol * fwdVol * dt;
-  impVol[0] = sqrt(var[0] / theTimes[0]);
+  impVol[0] = std::sqrt(var[0] / theTimes[0]);
 
   // Remaining time steps
   for (int j = 1; j < m; j++) {
     dt = theTimes[j] - theTimes[j - 1];
 
     // calculate log return standard deviation
-    double sum =
-        accumulate(thelogReturns[j].begin(), thelogReturns[j].end(), 0.0);
-    double mean = sum / n;
-    double sq_sum =
-        inner_product(thelogReturns[j].begin(), thelogReturns[j].end(),
-                      thelogReturns[j].begin(), 0.0);
-    double stDev = sqrt(sq_sum / n - mean * mean);
+    double sum2 =
+        std::accumulate(thelogReturns[j].begin(), thelogReturns[j].end(), 0.0);
+    double mean2 = sum2 / n;
+    double sq_sum2 =
+        std::inner_product(thelogReturns[j].begin(), thelogReturns[j].end(),
+                           thelogReturns[j].begin(), 0.0);
+    double stDev2 = std::sqrt(sq_sum2 / n - mean2 * mean2);
 
     // calculates implied volatility
-    double fwdVol = stDev * sqrt(freq);
+    double fwdVol2 = stDev2 * std::sqrt(freq);
 
-    var[j] = var[j - 1] + fwdVol * fwdVol * dt;
-    impVol[j] = sqrt(var[j] / theTimes[j]);
+    var[j] = var[j - 1] + fwdVol2 * fwdVol2 * dt;
+    impVol[j] = std::sqrt(var[j] / theTimes[j]);
   }
 
   return impVol;
