@@ -8,7 +8,9 @@ from .native import (  # New exports; Termstructure enums
     DayCounterType,
     ExerciseStyle,
     Heston,
-    HullWhite,
+    Heston,
+    LocalVol,
+    OptionType,
     LocalVol,
     OptionType,
     Sabr,
@@ -32,7 +34,8 @@ __all__ = [
     "Sabr",
     "Heston",
     "LocalVol",
-    "HullWhite",
+    "LocalVol",
+    "cdf_normal",
     "cdf_normal",
     "pdf_normal",
     "implied_vol",
