@@ -1,6 +1,8 @@
 import numpy as np
 
+import pytest
 import velesquant.native as m
+from velesquant.models import CMSSpreadModel
 
 
 def test_cms_spread_init_and_pricing():
@@ -43,7 +45,7 @@ def test_cms_spread_init_and_pricing():
     )
 
     # Test Spread Option
-    val = cs.spreadOption(0.01, 1.0, -1.0)
+    val = cs.spread_option(0.01, 1.0, -1.0)
     assert val > 0.0
 
 
@@ -83,6 +85,78 @@ def test_cms_spread_simulation():
         0.0,
     )
 
-    sim_res = cs.simulationCMSs()
+    sim_res = cs.simulate()
     assert isinstance(sim_res, list)
     assert len(sim_res) == 2
+
+
+def test_cms_spread_wrapper_init():
+    strikes = np.array([0.01, 0.02, 0.03]).reshape(-1, 1)
+    quotes = np.array([0.2, 0.2, 0.2]).reshape(-1, 1)
+    model = CMSSpreadModel(
+        1.0,
+        10.0,
+        0.03,
+        8.5,
+        1.0,
+        0.97,
+        0.85,
+        strikes,
+        quotes,
+        "Price",
+        1.0,
+        10.0,
+        0.02,
+        8.5,
+        1.0,
+        0.97,
+        0.85,
+        strikes,
+        quotes,
+        "Price",
+        0.5,
+    )
+    assert model._cpp_model is not None
+    assert model.spread_option(0.01) > 0
+
+    # Test to_dict
+    d = model.to_dict()
+    assert d["type"] == "CMSSpreadModel"
+    assert d["corr"] == 0.5
+
+    # Test simulate (no args in wrapper)
+    sim = model.simulate()
+    assert len(sim) == 2
+
+    # Test price stub
+    with pytest.raises(NotImplementedError):
+        model.price(None, None)
+
+    # Test Vol target
+    model_vol = CMSSpreadModel(
+        1.0,
+        10.0,
+        0.03,
+        1.0,
+        0.04,
+        1.0,
+        0.97,
+        strikes,
+        quotes,
+        "Vol",
+        1.0,
+        10.0,
+        0.02,
+        1.0,
+        0.04,
+        1.0,
+        0.97,
+        strikes,
+        quotes,
+        "Vol",
+        0.5,
+    )
+    assert model_vol._cpp_model is not None
+
+    with pytest.raises(NotImplementedError):
+        model.calibrate([], None)
