@@ -42,29 +42,33 @@ def test_heston_calibration_example():
     """
     Example of using the new Object-Oriented Heston API for calibration.
     """
-    # 1. Define Market Surface Data
-    # Heston requires a volatility surface (maturity x strike)
-    maturities = np.array([[1.0]])
-    forwards = np.array([[100.0]])
-    strikes = np.array([[100.0]])
-    quotes = np.array([[0.20]])
+    # Create "Ground Truth" to generate consistent data
+    true_kappa = 1.5
+    true_theta = 0.05
+    model_true = HestonModel(100.0, 0.04, true_kappa, true_theta, 0.1, -0.5)
 
-    # 2. Instantiate the Heston Model
+    # 1. Define Market Surface Data
+    maturities = np.array([1.0] * 9).reshape(-1, 1)
+    forwards = np.array([100.0] * 9).reshape(-1, 1)
+    strikes_list = [80.0, 85.0, 90.0, 95.0, 100.0, 105.0, 110.0, 115.0, 120.0]
+    strikes = np.array(strikes_list).reshape(-1, 1)
+
+    prices_list = [model_true.price_option(1.0, 100.0, k, "call") for k in strikes_list]
+    # Quotes are prices here because target="Price"
+    quotes = np.array(prices_list).reshape(-1, 1)
+
+    # 2. Instantiate the Heston Model (Initial Guess)
     heston = HestonModel(
-        spot=100.0, var0=0.04, kappa=1.0, theta=0.04, xi=0.1, rho=-0.5, seed=42
+        spot=100.0, var0=0.04, kappa=1.2, theta=0.04, xi=0.1, rho=-0.5, seed=42
     )
 
     # 3. Running Calibration
-
-    # Wrapper returns updated self OR params dict?
-    # Wrapper implementation returns self.
-
     result_model = heston.calibrate(
         maturities=maturities,
         forwards=forwards,
         strikes=strikes,
         quotes=quotes,
-        calibration_target="Price",  # Or "Volatility"
+        calibration_target="Price",
     )
 
     # 4. Verify Results
