@@ -1,4 +1,6 @@
+import pytest
 import velesquant.native as m
+from velesquant.models import LogNormalBasketModel
 
 
 def test_log_basket_init_and_sim():
@@ -21,11 +23,40 @@ def test_log_basket_init_and_sim():
 
     lb = m.LogNormalBasket(spots, strikes, maturities, forwards, iv, correlation)
 
-    assert lb.get_nassets() == 2
+    assert lb.get_n_assets() == 2
 
     schedule = [1.0]
-    paths = lb.simulate_basket(schedule)
+    paths = lb.simulate(schedule)
     # Returns Mdoub: [sim][asset]? Or [asset][sim]?
     # Assuming list of lists
     assert isinstance(paths, list)
     assert len(paths) > 0
+
+
+def test_basket_wrapper_init():
+    spots = [100.0, 110.0]
+    strikes = [100.0, 110.0]
+    maturities = [1.0, 2.0]
+    fwds = [[100.0, 100.0], [110.0, 110.0]]
+    ivs = [[0.2, 0.2], [0.22, 0.22]]
+    corr = [[1.0, 0.5], [0.5, 1.0]]
+    model = LogNormalBasketModel(spots, strikes, maturities, fwds, ivs, corr)
+    assert model.get_num_assets() == 2
+
+    # Test simulation variants
+    sim1 = model.simulate([1.0])
+    assert len(sim1) > 0
+    sim2 = model.simulate([1.0], with_rate_curves=True)
+    assert len(sim2) > 0
+
+    # Calibration stub test
+    with pytest.raises(NotImplementedError):
+        model.calibrate([], None)
+
+    with pytest.raises(NotImplementedError):
+        model.price(None, None)
+
+    # Test to_dict
+    d = model.to_dict()
+    assert d["type"] == "LogNormalBasketModel"
+    assert d["spots"] == spots
