@@ -1,10 +1,12 @@
-from velesquant import SchobelZhu
+import pytest
+from velesquant import native
+from velesquant.models import SchobelZhuModel
 
 
 def test_schobzhu_binding():
     # SchobelZhu(spot, var0, kappa, theta, xi, rho)
-    sz = SchobelZhu(100.0, 0.04, 1.0, 0.04, 0.3, -0.5)
-    price = sz.SchobelPrice(1.0, 100.0, 100.0)
+    sz = native.SchobelZhu(100.0, 0.04, 1.0, 0.04, 0.3, -0.5)
+    price = sz.price(1.0, 100.0, 100.0)
     assert isinstance(price, float)
     assert price > 0
 
@@ -16,11 +18,11 @@ def test_schobzhu_simulation():
     """
     Scenario: Monte Carlo simulation of Schobel-Zhu price paths.
     """
-    sz = SchobelZhu(100.0, 0.04, 1.0, 0.04, 0.3, -0.5)
+    sz = native.SchobelZhu(100.0, 0.04, 1.0, 0.04, 0.3, -0.5)
     times = [0.1, 0.5, 1.0]
     forwards = [100.0, 101.0, 102.0]
 
-    paths = sz.simulation(times, forwards)
+    paths = sz.simulate(times, forwards)
 
     assert len(paths) == len(times)
     # Check for no NaNs or Infs
@@ -33,17 +35,82 @@ def test_schobzhu_calibration():
     """
     Scenario: Calibrating Schobel-Zhu to a set of market quotes.
     """
-    sz = SchobelZhu(100.0, 0.04, 1.0, 0.04, 0.3, -0.5)
+    # 1. True Model
+    true_kappa = 1.0
+    true_theta = 0.04
+    true_var0 = 0.04
+    true_xi = 0.3
+    true_rho = -0.5
+    sz_true = native.SchobelZhu(
+        100.0, true_var0, true_kappa, true_theta, true_xi, true_rho
+    )
 
-    # Mock market data
-    maturities = [1.0, 1.0, 1.0]
-    forwards = [100.0, 100.0, 100.0]
-    strikes = [90.0, 100.0, 110.0]
-    quotes = [12.0, 5.0, 2.0]  # Rough call prices
+    # 2. Generate Synthetic Market Data
+    maturities = [1.0, 1.0, 1.0, 1.0, 1.0]
+    forwards = [100.0, 100.0, 100.0, 100.0, 100.0]
+    strikes = [80.0, 90.0, 100.0, 110.0, 120.0]
+    quotes = []
+    for k in strikes:
+        p = sz_true.price(1.0, 100.0, k)
+        quotes.append(p)
 
-    # Run calibration
-    sz.calibrator(maturities, forwards, strikes, quotes)
+    # 3. Guess Model (perturbed parameters - closer to truth)
+    sz_guess = native.SchobelZhu(100.0, 0.045, 0.9, 0.045, 0.25, -0.45)
+
+    # 4. Calibrate
+    sz_guess.calibrate(maturities, forwards, strikes, quotes)
+
+    # Check convergence (parameters should be closer to true)
+    # Note: Optimization might not recover exact values due to flat surface or local minima,
+    # but it should run without error and update params.
+    sz = sz_guess  # alias for assertions below
 
     # Check that parameters have changed (or at least are accessible)
     assert sz.var0 > 0
     assert abs(sz.rho) <= 1.0
+
+    # Use better initial guess for wrapper test too
+    model = SchobelZhuModel(100.0, 0.045, 0.9, 0.045, 0.25, -0.45)
+    assert model.spot == 100.0
+    assert model.kappa == 0.9
+    assert model.price_european(1.0, 100.0, 100.0) > 0
+
+    # Test calibration in wrapper
+    model.calibrate_to_quotes(
+        maturities=maturities,
+        forwards=forwards,
+        strikes=strikes,
+        quotes=quotes,
+        target="Price",
+    )
+    assert model.var0 > 0
+
+    # Test Properties
+    model.spot = 105.0
+    assert model.spot == 105.0
+    model.var0 = 0.05
+    assert model.var0 == 0.05
+    model.kappa = 2.5
+    assert model.kappa == 2.5
+    model.theta = 0.25
+    assert model.theta == 0.25
+    model.xi = 0.35
+    assert model.xi == 0.35
+    model.rho = -0.45
+    assert model.rho == -0.45
+
+    # Test to_dict
+    d = model.to_dict()
+    assert d["type"] == "SchobelZhuModel"
+    assert d["spot"] == 105.0
+    assert d["rho"] == -0.45
+
+    # Test simulate
+    sim = model.simulate([1.0], [105.0])
+    assert len(sim) > 0
+
+    # Test Not Implemented
+    with pytest.raises(NotImplementedError):
+        model.calibrate([], None)
+    with pytest.raises(NotImplementedError):
+        model.price(None, None)
