@@ -8,135 +8,149 @@ Verifies the newly exposed methods:
 """
 
 import pytest
-
 import velesquant.native as n
+
+
+def create_hwpde_model(
+    kappa, time_sigmas, sigmas, time_dfs, dfs, grid_points=100, time_step=0.01
+):
+    """Helper to create HWPDE with a HullWhiteModel."""
+    model = n.HullWhiteModel(kappa, time_sigmas, sigmas, time_dfs, dfs)
+    return n.HWPDE(model, grid_points, time_step)
 
 
 def test_hwpde_basic_pricing():
     """Test basic HWPDE pricing methods."""
-    # Create HWPDE with DF-based constructor
-    hwpde = n.HWPDE(
+    hwpde = create_hwpde_model(
         kappa=0.05,
-        timeSigmas=[1.0, 5.0, 10.0],
+        time_sigmas=[1.0, 5.0, 10.0],
         sigmas=[0.01, 0.012, 0.015],
-        timeDF=[0.0, 1.0, 5.0, 10.0, 30.0],
-        DF=[1.0, 0.95, 0.80, 0.65, 0.30],
+        time_dfs=[0.0, 1.0, 5.0, 10.0, 30.0],
+        dfs=[1.0, 0.95, 0.80, 0.65, 0.30],
+        grid_points=100,
+        time_step=0.01,
     )
 
     # Test pricingZB
-    zb_price = hwpde.pricingZB(5.0)
+    zb_price = hwpde.price_zero_bond(5.0)
     assert 0 < zb_price <= 1.0
 
     # Test pricingSwaption (already existed)
-    swaption = hwpde.pricingSwaption(1.0, 5.0, 0.03, 0.5)
+    swaption = hwpde.price_swaption(1.0, 5.0, 0.03, 0.5)
     assert swaption >= 0
 
     # Test pricingSwap
-    swap = hwpde.pricingSwap(0.0, 5.0, 0.03, 0.5)
+    swap = hwpde.price_swap(0.0, 5.0, 0.03, 0.5)
     assert isinstance(swap, float)
 
     # Test pricingZBO
-    zbo = hwpde.pricingZBO(1.0, 5.0, 0.90, "Call")
+    zbo = hwpde.price_zero_bond_option(1.0, 5.0, 0.90, n.OptionType.Call)
     assert zbo >= 0
 
     # Test pricingCouponBond
-    coupon_bond = hwpde.pricingCouponBond(0.0, 5.0, 0.04, 0.5)
+    coupon_bond = hwpde.price_coupon_bond(0.0, 5.0, 0.04, 0.5)
     assert coupon_bond > 0
 
     # Test pricingCBO
-    cbo = hwpde.pricingCBO(1.0, 5.0, 0.04, 1.0, 0.5, "Call")
+    cbo = hwpde.price_coupon_bond_option(1.0, 5.0, 0.04, 1.0, 0.5, n.OptionType.Call)
     assert cbo >= 0
 
 
 def test_hwpde_exotic_pricing():
     """Test exotic pricing methods."""
-    hwpde = n.HWPDE(
+    hwpde = create_hwpde_model(
         kappa=0.05,
-        timeSigmas=[1.0, 5.0],
+        time_sigmas=[1.0, 5.0],
         sigmas=[0.01, 0.012],
-        timeDF=[0.0, 1.0, 5.0, 10.0],
-        DF=[1.0, 0.95, 0.80, 0.65],
+        time_dfs=[0.0, 1.0, 5.0, 10.0],
+        dfs=[1.0, 0.95, 0.80, 0.65],
+        grid_points=100,
+        time_step=0.01,
     )
 
     # Test pricingBermudan
     exercises = [2.0, 3.0, 4.0]  # Exercise dates
-    bermudan = hwpde.pricingBermudan(1.0, 5.0, exercises, 0.03, 0.5)
+    bermudan = hwpde.price_bermudan(1.0, 5.0, exercises, 0.03, 0.5)
     assert bermudan >= 0
 
     # Test pricingCallableSwap
-    callable_swap = hwpde.pricingCallableSwap(
-        1.0, 5.0, exercises, 0.04, 1.0, 0.5, "Call"
+    callable_swap = hwpde.price_callable_swap(
+        1.0, 5.0, exercises, 0.04, 1.0, 0.5, n.OptionType.Call
     )
     assert isinstance(callable_swap, float)
 
 
 def test_hwpde_analysis():
     """Test analysis and getter methods."""
-    hwpde = n.HWPDE(
+    hwpde = create_hwpde_model(
         kappa=0.05,
-        timeSigmas=[1.0, 5.0],
+        time_sigmas=[1.0, 5.0],
         sigmas=[0.01, 0.012],
-        timeDF=[0.0, 1.0, 5.0, 10.0],
-        DF=[1.0, 0.95, 0.80, 0.65],
+        time_dfs=[0.0, 1.0, 5.0, 10.0],
+        dfs=[1.0, 0.95, 0.80, 0.65],
+        grid_points=100,
+        time_step=0.01,
     )
 
     # Test getSwapRate
-    swap_rate = hwpde.getSwapRate(1.0, 5.0, 0.5)
+    swap_rate = hwpde.get_swap_rate(1.0, 5.0, 0.5)
     assert -0.01 < swap_rate < 0.20
 
     # Test getImpVolATM
-    imp_vol = hwpde.getImpVolATM(1.0, 5.0, 0.5)
+    imp_vol = hwpde.get_implied_vol_atm(1.0, 5.0, 0.5)
     assert 0 < imp_vol < 2.0
 
     # Test getDFs
     time_points = [1.0, 2.0, 5.0]
-    dfs = hwpde.getDFs(time_points)
+    dfs = hwpde.get_discount_factors(time_points)
     assert len(dfs) == 3
     assert all(isinstance(x, float) for x in dfs)
 
     # Test simulationPDE
     sim_times = [0.5, 1.0, 2.0]
-    sim_path = hwpde.simulationPDE(sim_times)
+    sim_path = hwpde.simulate(sim_times)
     assert len(sim_path) == 3
     assert all(isinstance(x, float) for x in sim_path)
 
     # Test getters
-    assert hwpde.getKappa() == 0.05
-    assert hwpde.getTimeSigmas() == [1.0, 5.0]
-    assert hwpde.getSigmas() == [0.01, 0.012]
-    assert isinstance(hwpde.getR0(), float)
-    assert len(hwpde.getThetas()) > 0
+    assert hwpde.get_kappa() == 0.05
+    assert hwpde.get_time_sigmas() == [1.0, 5.0]
+    assert hwpde.get_sigmas() == [0.01, 0.012]
+    assert isinstance(hwpde.get_initial_rate(), float)
+    assert len(hwpde.get_thetas()) > 0
 
 
 def test_hwpde_calibration():
     """Test calibration method."""
-    hwpde = n.HWPDE(
+    hwpde = create_hwpde_model(
         kappa=0.05,
-        timeSigmas=[1.0, 5.0],
+        time_sigmas=[1.0, 5.0],
         sigmas=[0.01, 0.012],
-        timeDF=[0.0, 1.0, 5.0, 10.0],
-        DF=[1.0, 0.95, 0.80, 0.65],
+        time_dfs=[0.0, 1.0, 5.0, 10.0],
+        dfs=[1.0, 0.95, 0.80, 0.65],
+        grid_points=100,
+        time_step=0.01,
     )
 
     # Create swap quotes for calibration
     swaps = []
 
     s1 = n.DefSwap()
-    s1.Expiry = 1.0
-    s1.Tenor = 5.0
-    s1.Frequency = 0.5
-    s1.SwapRate = 0.03
-    s1.VolATM = 0.25
-    s1.Value = 0.0
+    s1.expiry = 1.0
+    s1.tenor = 5.0
+    s1.frequency = 0.5
+    s1.swap_rate = 0.03
+    s1.vol_atm = 0.25
+    s1.value = 0.0
     swaps.append(s1)
 
     s2 = n.DefSwap()
-    s2.Expiry = 5.0
-    s2.Tenor = 5.0
-    s2.Frequency = 0.5
-    s2.SwapRate = 0.04
-    s2.VolATM = 0.20
-    s2.Value = 0.0
+    s2.expiry = 5.0
+    s2.tenor = 5.0
+    s2.frequency = 0.5
+    s2.swap_rate = 0.04
+    s2.vol_atm = 0.20
+    s2.value = 0.0
     swaps.append(s2)
 
     time_dfs = [0.0, 1.0, 5.0, 10.0]
@@ -145,11 +159,10 @@ def test_hwpde_calibration():
     try:
         hwpde.calibrate(time_dfs, dfs, swaps)
         # Check that calibration changed parameters or finished
-        assert hwpde.getKappa() is not None
+        assert hwpde.get_kappa() is not None
     except RuntimeError:
         # We expect a RuntimeError with "too much freedom" or similar
         # because the input data is dummy data and might not converge/be valid
-        # This confirms the binding is reachable and throws the expected C++ exception
         pass
     except Exception as e:
         pytest.fail(f"Unexpected exception during calibration binding call: {e}")
