@@ -1,17 +1,29 @@
-from velesquant import AfSabr, AntonovSabr
+from velesquant import AfSabr, AntonovSabr, Sabr
 
 
 def test_sabr_pde_binding():
-    # AfSabr(alpha, beta, nu, rho, shift, maturity, F, sizeX, sizeT, nd)
-    af = AfSabr(0.2, 0.5, 0.3, -0.4, 0.0, 1.0, 0.03, 100, 100, 3.0)
+    # Setup Sabr model
+    # Sabr(maturity, forward, beta, alpha, nu, rho, shift)
+    T = 1.0
+    F = 0.03
+    alpha = 0.2
+    beta = 0.5
+    nu = 0.3
+    rho = -0.4
+    shift = 0.0
+
+    model = Sabr(T, F, beta, alpha, nu, rho, shift)
+
+    # AfSabr(model, sizeX, sizeT, nd)
+    af = AfSabr(model, 100, 100, 3.0)
 
     # Check base class methods
-    assert af.getAlpha() == 0.2
-    assert af.getBeta() == 0.5
+    assert af.get_alpha() == 0.2
+    assert af.get_beta() == 0.5
 
     # AntonovSabr
-    ant = AntonovSabr(0.2, 0.5, 0.3, -0.4, 1.0, 0.03, 100, 100, 3.0)
-    assert ant.getNu() == 0.3
+    ant = AntonovSabr(model, 100, 100, 3.0)
+    assert ant.get_nu() == 0.3
 
 
 def test_sabr_pde_density():
@@ -25,25 +37,42 @@ def test_sabr_pde_density():
     beta = 0.5
     nu = 0.3
     rho = -0.4
+    shift = 0.0
+
+    model = Sabr(T, F, beta, alpha, nu, rho, shift)
 
     # Increase number of standard deviations for grid boundary to capture tails
     # and increase grid points for better integration accuracy
-    sabr = AfSabr(alpha, beta, nu, rho, 0.0, T, F, 500, 100, 8.0)
+    sabr = AfSabr(model, 500, 100, 8.0)
 
     # Get density
-    density = sabr.getDensity()
-    f_grid = sabr.getFgrid()
+    density = sabr.get_density()
+    f_grid = sabr.get_f_grid()
 
     assert len(density) == 500
     assert len(f_grid) == 500
 
     # Simple trapezoidal integration to check if it integrates to ~1
-    integral = 0.0
-    for i in range(len(density) - 1):
-        h = f_grid[i + 1] - f_grid[i]
-        avg_h = (density[i] + density[i + 1]) / 2.0
-        integral += avg_h * h
+    def integrate(d, f):
+        res = 0.0
+        for i in range(len(d) - 1):
+            h = f[i + 1] - f[i]
+            avg_h = (d[i] + d[i + 1]) / 2.0
+            res += avg_h * h
+        return res
 
-    # It might not be exactly 1 depending on model specifics/normalization or if result is not PDF.
-    # Just verify we got a non-trivial distribution.
-    assert integral > 0.0
+    integral_af = integrate(density, f_grid)
+
+    # AfSabr might have absorbing boundary issues or truncation, so we check > 0.5
+    # (Investigated: ~0.74 with current parameters)
+    assert integral_af > 0.5
+
+    # Check AntonovSabr (should be very close to 1.0)
+    ant = AntonovSabr(model, 500, 100, 8.0)
+    density_ant = ant.get_density()
+    f_grid_ant = ant.get_f_grid()
+    assert len(density_ant) == 500
+
+    integral_ant = integrate(density_ant, f_grid_ant)
+    # Be slightly lenient for numerical integration errors
+    assert abs(integral_ant - 1.0) < 0.01
