@@ -1,8 +1,11 @@
 #include <cmath>
 #include <gtest/gtest.h>
 #include <iostream>
+#include <memory>
 #include <vector>
-#include <velesquant/models/hw.h>
+#include <velesquant/engines/hullwhite_analytic_engine.h>
+#include <velesquant/models/hullwhite_model.h>
+#include <velesquant/models/utility.h> // for annuity
 
 using namespace velesquant;
 
@@ -16,6 +19,29 @@ std::vector<double> generateDFs(const std::vector<double> &times, double rate) {
   }
   return dfs;
 }
+
+// Calculate swap rate using model
+double calculateSwapRate(const models::HullWhiteModel &model, double expiry,
+                         double tenor, double payFreq) {
+  double startTime = expiry;
+  double endTime = expiry + tenor;
+
+  double pStart = model.getDiscountFactor(startTime);
+  double pEnd = model.getDiscountFactor(endTime);
+
+  double annuity = 0.0;
+  int nPeriods = static_cast<int>(std::round(tenor / payFreq));
+
+  for (int i = 1; i <= nPeriods; ++i) {
+    double t = startTime + i * payFreq;
+    annuity += payFreq * model.getDiscountFactor(t);
+  }
+
+  if (annuity < 1e-12)
+    return 0.0;
+  return (pStart - pEnd) / annuity;
+}
+
 } // namespace
 
 class HullWhiteTestFixture : public ::testing::Test {
@@ -38,38 +64,32 @@ protected:
 };
 
 TEST_F(HullWhiteTestFixture, Construction) {
-  HullWhite hw(kappa, timeSigmas, sigmas, timeDFs, DFs);
-  EXPECT_NEAR(hw.getKappa(), kappa, 1e-12);
-  auto gotSigmas = hw.getSigmas();
+  auto model = std::make_shared<models::HullWhiteModel>(kappa, timeSigmas,
+                                                        sigmas, timeDFs, DFs);
+  EXPECT_NEAR(model->getKappa(), kappa, 1e-12);
+  auto gotSigmas = model->getSigmas();
   EXPECT_EQ(gotSigmas.size(), 2);
   EXPECT_DOUBLE_EQ(gotSigmas[0], sigma);
 }
 
 TEST_F(HullWhiteTestFixture, ZCBondPricing) {
-  HullWhite hw(kappa, timeSigmas, sigmas, timeDFs, DFs);
+  auto model = std::make_shared<models::HullWhiteModel>(kappa, timeSigmas,
+                                                        sigmas, timeDFs, DFs);
 
   // Price a bond maturing in 1 year
-  // For T=1, ZC bond price should be close to DF(1) = exp(-0.05) ~ 0.9512
-  double price = hw.ZC(1.0);
+  // For T=1, ZC bond price should be exactly DF(1) = exp(-0.05) ~ 0.9512
+  double price = model->getDiscountFactor(1.0);
   double expected = DFs[1]; // timeDFs[1] is 1.0
 
-  // Verify ZC consistency with its own implementation (approximating expected
-  // behavior) Current implementation of ZC in hw.cpp applies an extra discount
-  // factor exp(-B * r0) effectively double discounting if r0 comes from the
-  // curve. We strictly test that it returns a valid price (0 < p < 1)
-  EXPECT_GT(price, 0.0);
-  EXPECT_LT(price, 1.0);
-
-  // Check against formula explicitly to ensure code hasn't changed
-  // B(0, 1) approx (1 - exp(-0.1))/0.1 = 0.9516
-  // Factor = exp(-0.9516 * (0.05 + small_var)) ~ exp(-0.047) ~ 0.954
-  // P_model ~ 0.9512 * 0.954 ~ 0.907
-  // The failing value was 0.90699... so this hypothesis holds.
-  EXPECT_NEAR(price, 0.907, 1e-3);
+  // The ZC(T) should be exactly the discount factor from the curve
+  EXPECT_NEAR(price, expected, 1e-12);
 }
 
 TEST_F(HullWhiteTestFixture, OptionOnBond) {
-  HullWhite hw(kappa, timeSigmas, sigmas, timeDFs, DFs);
+  auto model = std::make_shared<models::HullWhiteModel>(kappa, timeSigmas,
+                                                        sigmas, timeDFs, DFs);
+  auto engine = std::make_shared<
+      engines::HullWhiteAnalyticEngine<models::HullWhiteModel>>(model);
 
   double expiry = 1.0;
   double maturity = 2.0;
@@ -78,11 +98,12 @@ TEST_F(HullWhiteTestFixture, OptionOnBond) {
   double forwardBondPrice = DFs[2] / DFs[1];
   double strike = forwardBondPrice;
 
-  double callPrice = hw.optionBond(expiry, maturity, strike, OptionType::Call);
-  double putPrice = hw.optionBond(expiry, maturity, strike, OptionType::Put);
+  double callPrice =
+      engine->optionBond(expiry, maturity, strike, OptionType::Call);
+  double putPrice =
+      engine->optionBond(expiry, maturity, strike, OptionType::Put);
 
-  // Put-Call Parity for Bond Option in this model (using Market DFs as
-  // optionBond does): C - P = P(0, M) - K * P(0, E)
+  // Put-Call Parity: C - P = P(0, M) - K * P(0, E)
   double parityTarget = DFs[2] - strike * DFs[1];
 
   // Since strike = DFs[2]/DFs[1], parityTarget should be 0.
@@ -92,30 +113,34 @@ TEST_F(HullWhiteTestFixture, OptionOnBond) {
 }
 
 TEST_F(HullWhiteTestFixture, SwaptionPricing) {
-  HullWhite hw(kappa, timeSigmas, sigmas, timeDFs, DFs);
+  auto model = std::make_shared<models::HullWhiteModel>(kappa, timeSigmas,
+                                                        sigmas, timeDFs, DFs);
+  auto engine = std::make_shared<
+      engines::HullWhiteAnalyticEngine<models::HullWhiteModel>>(model);
 
   double expiry = 1.0;
   double tenor = 2.0;
   double payFreq = 0.5; // Semi-annual
 
   // Get forward swap rate
-  double swapRate = hw.getSwapRate(expiry, tenor, payFreq);
+  double swapRate = calculateSwapRate(*model, expiry, tenor, payFreq);
   EXPECT_GT(swapRate, 0.0);
   EXPECT_LT(swapRate, 0.2); // Reasonable swap rate
 
   // Price ATM swaption
   double strike = swapRate;
-  double price = hw.swaption(expiry, tenor, strike, payFreq);
+  double price = engine->swaption(expiry, tenor, strike, payFreq);
 
   EXPECT_GT(price, 0.0);
   EXPECT_LT(price, 1.0); // Swaption price should be reasonable
 }
 
 TEST_F(HullWhiteTestFixture, Simulation) {
-  HullWhite hw(kappa, timeSigmas, sigmas, timeDFs, DFs);
+  auto model = std::make_shared<models::HullWhiteModel>(kappa, timeSigmas,
+                                                        sigmas, timeDFs, DFs);
 
   std::vector<double> times = {0.25, 0.5, 0.75, 1.0};
-  std::vector<double> rates = hw.simulation(times);
+  std::vector<double> rates = model->simulation(times);
 
   EXPECT_EQ(rates.size(), times.size());
 
